Make TraitRenamedCaseOnly extend Delta
