+++ src/PHPSemVerChecker/Operation/TraitRenamedCaseOnly.php
@@ -5,7 +5,7 @@ namespace PHPSemVerChecker\Operation;
 use PhpParser\Node\Stmt\Trait_;
 use PHPSemVerChecker\Node\Statement\Trait_ as PTrait;
 
-class TraitRenamedCaseOnly extends Operation {
+class TraitRenamedCaseOnly extends TraitOperationDelta {
 	/**
 	 * @var string
 	 */
@@ -14,46 +14,4 @@ class TraitRenamedCaseOnly extends Operation {
 	 * @var string
 	 */
 	protected $reason = 'Trait was renamed (case only).';
-	/**
-	 * @var string
-	 */
-	protected $fileAfter;
-	/**
-	 * @var \PhpParser\Node\Stmt\Trait_
-	 */
-	protected $traitAfter;
-
-	/**
-	 * @param string                      $fileAfter
-	 * @param \PhpParser\Node\Stmt\Trait_ $traitAfter
-	 */
-	public function __construct($fileAfter, Trait_ $traitAfter)
-	{
-		$this->fileAfter = $fileAfter;
-		$this->traitAfter = $traitAfter;
-	}
-
-	/**
-	 * @return string
-	 */
-	public function getLocation()
-	{
-		return $this->fileAfter;
-	}
-
-	/**
-	 * @return int
-	 */
-	public function getLine()
-	{
-		return $this->traitAfter->getLine();
-	}
-
-	/**
-	 * @return string
-	 */
-	public function getTarget()
-	{
-		return PTrait::getFullyQualifiedName($this->traitAfter);
-	}
 }
